Rewrote Vehicle Cargo script
The old legacy version is still included, but it's no longer active.

The new version is called vehicle_pickup2, and it's the one that's enabled by default.

More testing is still needed. But it seems much more promising right now. Which is why I made this the new default.

Nothing should have changed in user experience, it should still act the same. With the one exception it might feel a bit smoother.

Code wise, a lot has changed. And it's way more readable, and streamlined. Making it easier to update.

Such as, now vehicles can have more than one sphere of influences.

Which means, trucks are properly covered from front to back now.

As well, as the Interior of Pelicans can also carry cargo now.

The Elephant even have up to three sphere of influences. There is virtually no upper limit. But the more we have, the more likely it's going to cause lag.

But I did not notice any lag whatsoever with testing in singleplayer.

Which is a good sign.

So will have to check other vehivles too in the future, and see if I can adjust their sphere of infleunces too. And make it make more sense.